Time: 2 ms (45.90%), Space: 38.4 MB (11.06%) - LeetHub

+++ 69-sqrtx/69-sqrtx.java
@@ -1,14 +1,14 @@
 class Solution {
     public int mySqrt(int x) {
-        long start = 0;
-        long end = x;
+        int start = 0;
+        int end = x;
         
         while(start<=end){
             long mid = start + (end - start)/2;
-            if(mid * mid > x) end = mid-1;
-            else start = mid+1;
+            if(mid * mid > x) end = (int)mid-1;
+            else start = (int)mid+1;
         }
         
-    return (int)end;
+    return end;
     }
 }
